backend/requirements.txt
fastapi==0.70.0
uvicorn[standard]==0.15.0
pdfplumber==0.7.1
requests==2.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
email-validator==1.2.1
pyotp==2.7.0
loguru==0.6.0
slowapi==0.1.7
pydantic[email]==1.9.0
pyOpenSSL==22.0.0
py-bcrypt==0.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pyotp==2.9.0
pydantic-extra-types==2.7.1
sqlite-utils==3.35
loguru==0.7.2
numpy==1.26.4
pandas==2.2.1
matplotlib==3.8.3
